# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/replies/my-replies": {
            "get": {
                "tags": [
                    "Show User's Replies"
                ],
                "summary": "Show User's Replies",
                "description": "Show User's Replies",
                "operationId": "show user's Replies",
                "responses": {
                    "200": {
                        "description": "successful updated post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/replies/create": {
            "post": {
                "tags": [
                    "Add new Replies"
                ],
                "summary": "Add new Replies",
                "description": "Add new Replies",
                "operationId": "add new replies",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "text",
                                    "comment_id"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "text"
                                    },
                                    "comment_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful created comment",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/replies/update/{reply}": {
            "put": {
                "tags": [
                    "Update User Replies"
                ],
                "summary": "Update User Replies",
                "description": "Update User Replies",
                "operationId": "update user replies",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID of the reply to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful updated Replies",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/replies/delete/{reply}": {
            "delete": {
                "tags": [
                    "Delete User Replies"
                ],
                "summary": "Delete User Replies",
                "description": "Delete User Replies",
                "operationId": "delete user replies",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID of the replies to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful Deleted Replies",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}